CIRC-1933 Implementation of search slips API
