+++ ramls/staff-slips.raml
@@ -1,5 +1,5 @@
 #%RAML 1.0
-title: Stuff Slips
+title: Staff Slips
 version: v0.3
 protocols: [ HTTP, HTTPS ]
 baseUri: http://localhost:9130
@@ -9,7 +9,7 @@ documentation:
     content: <b>API for staff slips generation</b>
 
 types:
-  stuff-slips: !include staff-slips-response.json
+  staff-slips: !include staff-slips-response.json
 
 traits:
   language: !include raml-util/traits/language.raml
@@ -23,10 +23,10 @@ resourceTypes:
         type:
           collection-get:
             exampleCollection: !include examples/staff-slips-response.json
-            schemaCollection: stuff-slips
+            schemaCollection: staff-slips
     /search-slips:
       /{servicePointId}:
         type:
           collection-get:
             exampleCollection: !include examples/staff-slips-response.json
-            schemaCollection: stuff-slips
+            schemaCollection: staff-slips
