+++ src/main/java/org/folio/circulation/domain/RequestTypeItemStatusWhiteList.java
@@ -25,6 +25,7 @@ import static org.folio.circulation.domain.ItemStatus.WITHDRAWN;
 import java.util.Arrays;
 import java.util.EnumMap;
 import java.util.List;
+import java.util.Map;
 
 public class RequestTypeItemStatusWhiteList {
   private static EnumMap<ItemStatus, Boolean> recallRules;
@@ -159,6 +160,15 @@ public class RequestTypeItemStatusWhiteList {
   public static List<RequestType> getRequestTypesAllowedForItemStatus(ItemStatus itemStatus) {
     return Arrays.stream(RequestType.values())
       .filter(requestType -> requestsRulesMap.get(requestType).get(itemStatus))
+      .toList();
+  }
+
+  public static List<ItemStatus> getItemStatusesAllowedForRequestType(RequestType requestType) {
+    return requestsRulesMap.get(requestType)
+      .entrySet()
+      .stream()
+      .filter(entry -> Boolean.TRUE.equals(entry.getValue()))
+      .map(Map.Entry::getKey)
       .toList();
   }
 }
